+++ collect_app/src/main/java/org/odk/collect/android/widgets/ExVideoWidget.java
@@ -12,12 +12,12 @@ import org.javarosa.core.model.data.StringData;
 import org.javarosa.form.api.FormEntryPrompt;
 import org.odk.collect.android.R;
 import org.odk.collect.android.application.Collect;
-import org.odk.collect.android.dao.helpers.ContentResolverHelper;
 import org.odk.collect.android.databinding.ExVideoWidgetAnswerBinding;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
 import org.odk.collect.android.utilities.ActivityAvailability;
 import org.odk.collect.android.utilities.ApplicationConstants;
 import org.odk.collect.android.utilities.ExternalAppIntentProvider;
+import org.odk.collect.android.utilities.FileUtils;
 import org.odk.collect.android.utilities.MediaUtils;
 import org.odk.collect.android.utilities.QuestionMediaManager;
 import org.odk.collect.android.utilities.ToastUtils;
@@ -106,7 +106,7 @@ public class ExVideoWidget extends QuestionWidget implements FileWidget, WidgetD
             if (object instanceof File) {
                 ToastUtils.showLongToast(R.string.invalid_file_type);
                 mediaUtils.deleteMediaFile(((File) object).getAbsolutePath());
-                Timber.e("ExVideoWidget's setBinaryData must receive a video file but received: %s", ContentResolverHelper.getMimeType((File) object));
+                Timber.e("ExVideoWidget's setBinaryData must receive a video file but received: %s", FileUtils.getMimeType((File) object));
             } else {
                 Timber.e("ExVideoWidget's setBinaryData must receive a video file but received: %s", object.getClass());
             }
